refactor: GuideSection 컴포넌트 분리화 작업
